+++ gradle-testkit-ext/main/kotlin/com/javiersc/gradle/testkit/extensions/Sandbox.kt
@@ -1,3 +1,5 @@
+@file:OptIn(ExperimentalStdlibApi::class)
+
 package com.javiersc.gradle.testkit.extensions
 
 import java.io.File
@@ -6,13 +8,16 @@ import java.nio.file.Path
 import java.nio.file.Paths
 import org.gradle.testkit.runner.GradleRunner
 
+public fun resourceFile(resource: String): File =
+    File(Thread.currentThread().contextClassLoader?.getResource(resource)?.toURI()!!)
+
 internal val sandboxPath: Path = Paths.get("build/sandbox").apply { toFile().mkdirs() }
 
 internal val sandboxIsolatedPath: Path =
     Paths.get("build/sandbox-isolated").apply { toFile().mkdirs() }
 
 internal fun String.copyResourceTo(destination: File) {
-    getResource(this).copyRecursively(destination)
+    resourceFile(this).copyRecursively(destination)
 }
 
 internal fun createSandboxDirectory(prefix: String): File =
@@ -23,10 +28,40 @@ internal fun createSandboxIsolatedDirectory(prefix: String): File =
 
 internal val GradleRunner.argumentsTxt: List<String>
     get() {
-        val argumentsFile = File("$projectDir/ARGUMENTS.txt")
-        return if (argumentsFile.exists()) argumentsFile.readLines().first().split(" ")
-        else emptyList()
+        val argumentsFile = projectDir.resolve("ARGUMENTS.txt")
+        return if (argumentsFile.exists()) {
+            buildList {
+                val line = argumentsFile.readLines().first()
+                var (nextArgument, remaining) = line.nextArgument()
+                nextArgument?.let(::add)
+                while (nextArgument != null || remaining != null) {
+                    val (nextArgumentWhile, remainingWhile) =
+                        remaining?.nextArgument() ?: (null to null)
+                    nextArgument = nextArgumentWhile
+                    remaining = remainingWhile
+                    nextArgument?.let(::add)
+                }
+            }
+        } else {
+            emptyList()
+        }
     }
 
-private fun getResource(resource: String): File =
-    File(Thread.currentThread().contextClassLoader?.getResource(resource)?.toURI()!!)
+private fun String.nextArgument(): Pair<String?, String?> {
+    val line = dropWhile(Char::isWhitespace)
+    val (nextArgument, remaining) =
+        when {
+            line.isBlank() -> null to null
+            line.startsWith("\"") -> {
+                val nextArgument = line.drop(1).takeWhile { it != '"' }
+                val remaining = line.drop(1 + nextArgument.length + 1)
+                nextArgument to remaining
+            }
+            else -> {
+                val nextArgument = line.takeWhile { !it.isWhitespace() }
+                val remaining = line.drop(nextArgument.length)
+                nextArgument to remaining
+            }
+        }
+    return nextArgument to remaining
+}