+++ gradle-testkit-ext/main/kotlin/com/javiersc/gradle/testkit/extensions/GradleTestKitExtensions.kt
@@ -81,9 +81,8 @@ public fun GradleRunner.withArgumentsFromTXT() {
     withArguments(argumentsTxt)
 }
 
-public fun GradleRunner.gradlewArgumentFromTXT() {
+public fun GradleRunner.gradlewArgumentFromTXT(): BuildResult =
     gradlew(*argumentsTxt.toTypedArray())
-}
 
 public fun GradleRunner.gradlew(vararg arguments: String): BuildResult {
     val previousArgs: List<String> = this.arguments.toList()